[KYUUBI #2694] EngineEvent.toString outputs application tags
### _Why are the changes needed?_
close #2694

Now `engineRefId` is part of the application name and is output in the engine log, such as SparkBatch, or some UTs cover the app name. At this time, it is difficult to correlate the engine log.
Now `engineRefId` is also part of tags, so we can output tags when outputting logs.

### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [ ] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #2695 from cxzl25/KYUUBI-2694.

Closes #2694

ae995a77 [sychen] output tags

Authored-by: sychen <sychen@ctrip.com>
Signed-off-by: ulysses-you <ulyssesyou@apache.org>