[KYUUBI #2924] Correct the frontend server start state
### _Why are the changes needed?_

CI has a lot of flaky test, it is because we will mark server is started when we call thread.start, but the server (bianry or http) may not be work at that moment. As a result, the client can not connect a started server.

### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [ ] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #2925 from ulysses-you/correct-start-state.

Closes #2924

b5a30696 [ulysses-you] test
bb4b4a8a [ulysses-you] sleep

Authored-by: ulysses-you <ulyssesyou18@gmail.com>
Signed-off-by: ulysses-you <ulyssesyou@apache.org>