[KYUUBI #3065] Support to retry the killApplicationByTag for JpsApplicationOperation
If using jps application manager, for spark job, it might generate two processes.

And the first one is only for spark submit args parser and is transient, so if we try to kill the app by tag, it might say `No such process` and the spark main process is still running.
So, for JpsApplicationOperation, we need to retry the kill app operation to prevent flaky test.

- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #3065 from turboFei/jps_appmgr.

Closes #3065

3d7503d7 [Fei Wang] thanks
7b259658 [Fei Wang] if else
9133e80d [Fei Wang] Retry the kill for one time

Authored-by: Fei Wang <fwang12@ebay.com>
Signed-off-by: Fei Wang <fwang12@ebay.com>