[KYUUBI #4025] Tidy up the `Are you willing to submit PR` option in the issue template
### _Why are the changes needed?_

This pr aims to tidy up the `Are you willing to submit PR` option in the issue template.

Currently, The `Are you willing to submit PR` option of some Apache kyuubi issue templates has three options as follow:
 - [ ] Yes. I can submit a PR independently to improve.
 - [ ] Yes. I would be willing to submit a PR with guidance from the Kyuubi community to improve.
 - [ ] No. I cannot submit a PR at this time.

The intent of `Are you willing to submit PR` option is to confirm whether the questioner can submit PR at now, So two options are enough.

### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [ ] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #4025 from Yikf/issue-template.

Closes #4025

11b08ee8 [Yikf] Tidy up the Are you willing to submit PR option in the issue template

Authored-by: Yikf <yikaifei1@gmail.com>
Signed-off-by: Cheng Pan <chengpan@apache.org>