[KYUUBI #4526][FOLLOWUP] Defer evaluation for GroupProvider
As described in the discussion: https://github.com/apache/kyuubi/discussions/4512

Since the pr https://github.com/apache/kyuubi/pull/3897 , we introduced the feature of `Supplying pluggable GroupProvider`.

But now when we create a new EngineRef, no matter which engine share level is used, getGroupNames will always be called.

This may cause users to encounter some confusion stack trace when they do not use the GROUP engine share level.

So we solve the problem through moving the `sessionManager.groupProvider.primaryGroup` to the inside of `EngineRef.appUser`, to ensure that the feature of `Supplying pluggable GroupProvider` is only valid for the GROUP engine share level.

- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.readthedocs.io/en/master/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #4527 from merrily01/branch-1.7-groupProvider.

Closes #4526

0a858a921 [maruilei] Address comments.
c63a0f97b [maruilei] [KYUUBI #4526][FOLLOWUP] Ensure that the feature of `Supplying pluggable GroupProvider` is only valid for the GROUP engine share level.

Authored-by: maruilei <maruilei@58.com>
Signed-off-by: Cheng Pan <chengpan@apache.org>