[KYUUBI #5206] Try to kill pod with label if no ApplicationInfo found to prevent pod leak
### _Why are the changes needed?_

Now for `KubernetesApplicationOperation`, it rely on the appInfoStore.

For batch rest api, if the closeBatch request can not be send to the kyuubiInstance that created the batch, the current kyuubiInstance will try to kill the batch.

I wonder that, in this case, the applicationInfo might can not be found in the appInfoStore.

It is better to try the best to kill the pod with  `kyuubi-unique-tag` label to prevent pod leak.

### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.readthedocs.io/en/master/contributing/code/testing.html#running-tests) locally before make a pull request

### _Was this patch authored or co-authored using generative AI tooling?_

Closes #5206 from turboFei/k8s_pod_kill.

Closes #5206

630bcd04b [fwang12] warning
a9c22e03f [fwang12] delete with label

Authored-by: fwang12 <fwang12@ebay.com>
Signed-off-by: fwang12 <fwang12@ebay.com>