[KYUUBI #5750] [Spark] Fix the inaccurate issue of obtaining COLUMN_SIZE in the decimal field jdbc of spark engine
# :mag: Description
## Issue References 🔗

This pull request fixes #

## Describe Your Solution 🔧

Fix the inaccuracy in getting COLUMN_SIZE from spark engine's decimal field jdbc. The current jdbc's get column size gets the decimal field as default size, which is inaccurate; if it is decimal(20,3), the obtained column size is 16; the actual is 20.

## Types of changes :bookmark:

- [X] Bugfix (non-breaking change which fixes an issue)

## Test Plan 🧪

#### Behavior Without This Pull Request :coffin:

#### Behavior With This Pull Request :tada:

#### Related Unit Tests

---

Closes #5750 from waywtdcc/support_spark_decimal2.

Closes #5750

2d288f5aa [waywtdcc] [Spark] Fix the inaccurate issue of obtaining COLUMN_SIZE in the decimal field jdbc of spark engine
4286354c2 [waywtdcc] Support flink engine under the select statement, the results can be read in a stream
e5b74b0c8 [waywtdcc] Support flink engine under the select statement, the results can be read in a stream

Authored-by: waywtdcc <waywtdcc@163.com>
Signed-off-by: Kent Yao <yao@apache.org>